Feature/codebase cleanup (#5)
* added comments, removed unused interfaces in diff.ts

* added comments in chat view provider

* removed unused functions in originalContentProvider.ts, added comments

* modified logging in messageHandlerService

* removed unused commands in package.json

* added diff toggle command in extension.ts

* added diff toggle in status bar and diff functionality

* removed quick diff implementation, added persistence of ai modified file diffs across sessions

* removed auto opening of diff view, now opens as separate tab, removed unused methods

* refactored contextService, removed unused functions, fixed unecessary re-render bug in diffService

* refactored fileModificationService, removed unused functions

* refactored mcp related code into separate services, extracted interfaces

* refactored context related services

* extracted promptService and modelMessageService, grouped services into folders

* added application of changes logic to fileModificationService

* removed unused code and interfaces

* extracted chatHistoryService from messageHandlerService

* updated README.md

* modified README.md

* modified folder structure to use domain driven design
