+++ src/features/diff/originalContentProvider.ts
