+++ src/types/diff.ts
@@ -1,4 +1,3 @@
-
 import * as vscode from 'vscode';
 
 /**
@@ -9,3 +8,13 @@ export interface TrackedFile {
   readonly originalUri: vscode.Uri;
   readonly originalContent: string;
 }
+
+/**
+ * Minimal payload that is stored in VS Code workspaceState to restore tracked files
+ * across sessions. The actual diff service rebuilds the virtual documents from this snapshot.
+ */
+export interface PersistedTrackedFile {
+  readonly filePath: string;
+  readonly originalContent: string;
+}
+
