+++ src/types/fileModification.ts
@@ -1,0 +1,9 @@
+/**
+ * Interface for file modification data
+ */
+export interface FileModification {
+  fileName: string;
+  filePath: string;
+  content: string;
+  summary?: string;
+}
